+++ pymcuprog/deviceinfo/devices/avr64ec28.py
@@ -1,14 +1,19 @@
+
+"""
+Required device info for the avr64ec28 devices
+"""
+
 from pymcuprog.deviceinfo.eraseflags import ChiperaseEffect
 
 DEVICE_INFO = {
     'interface': 'UPDI',
-    'name': 'avr16eb14',
+    'name': 'avr64ec28',
     'architecture': 'avr8x',
 
     # eeprom
-    'eeprom_address_byte': 0x1400,
+    'eeprom_address_byte': 0x00001400,
     'eeprom_size_bytes': 0x200,
-    'eeprom_page_size_bytes': 0x8,
+    'eeprom_page_size_bytes': 0x1,
     'eeprom_read_size_bytes': 0x01,
     'eeprom_write_size_bytes': 0x01,
     'eeprom_chiperase_effect': ChiperaseEffect.CONDITIONALLY_ERASED_AVR,
@@ -16,10 +21,10 @@ DEVICE_INFO = {
 
     # flash
     'flash_address_byte': 0x00800000,
-    'flash_size_bytes': 0x4000,
-    'flash_page_size_bytes': 0x40,
+    'flash_size_bytes': 0x10000,
+    'flash_page_size_bytes': 0x80,
     'flash_read_size_bytes': 0x02,
-    'flash_write_size_bytes': 0x40,
+    'flash_write_size_bytes': 0x80,
     'flash_chiperase_effect': ChiperaseEffect.ALWAYS_ERASED,
     'flash_isolated_erase': True,
 
@@ -33,8 +38,8 @@ DEVICE_INFO = {
     'fuses_isolated_erase': False,
 
     # internal_sram
-    'internal_sram_address_byte': 0x7800,
-    'internal_sram_size_bytes': 0x800,
+    'internal_sram_address_byte': 0x6800,
+    'internal_sram_size_bytes': 0x1800,
     'internal_sram_page_size_bytes': 0x01,
     'internal_sram_read_size_bytes': 0x01,
     'internal_sram_write_size_bytes': 0x01,
@@ -52,7 +57,7 @@ DEVICE_INFO = {
 
     # signatures
     'signatures_address_byte': 0x1080,
-    'signatures_size_bytes': 0x40,
+    'signatures_size_bytes': 0x80,
     'signatures_page_size_bytes': 0x01,
     'signatures_read_size_bytes': 0x01,
     'signatures_write_size_bytes': 0x00,
@@ -64,16 +69,16 @@ DEVICE_INFO = {
     'boot_row_size_bytes': 0x40,
     'boot_row_page_size_bytes': 0x40,
     'boot_row_read_size_bytes': 0x01,
-    'boot_row_write_size_bytes': 0x01,
+    'boot_row_write_size_bytes': 0x40,
     'boot_row_chiperase_effect': ChiperaseEffect.CONDITIONALLY_ERASED_AVR,
-    'boot_row_isolated_erase': True,    
+    'boot_row_isolated_erase': True,
 
     # user_row
     'user_row_address_byte': 0x1200,
     'user_row_size_bytes': 0x40,
     'user_row_page_size_bytes': 0x40,
     'user_row_read_size_bytes': 0x01,
-    'user_row_write_size_bytes': 0x01,
+    'user_row_write_size_bytes': 0x40,
     'user_row_chiperase_effect': ChiperaseEffect.NOT_ERASED,
     'user_row_isolated_erase': True,
 
@@ -84,5 +89,5 @@ DEVICE_INFO = {
     'address_size': '24-bit',
     'prog_clock_khz': 1800,
     'hv_implementation': 2,
-    'device_id': 0x1E9449,
+    'device_id': 0x1E962A,
 }